requirements.txt
pyrogram
Tgcrypto
python-decouple==3.6
 
